Update to functions_basicsII.py first_plus_length
Updated the code for the section first_plus_length where code is supposed to take first number of list and add it to the length of the list. The original code too the first number in the list and printed it and then printed the length of the list. Now it prints the first number added to the length of the list.